Some fix to formats layer (mainly import method) And rename formats/{spdx, cyclonedx}/models.py to document.py
